+++ lib/core/constants/strings.dart
@@ -1,5 +1,5 @@
 class Strings {
   // App
   static const appTitle = 'Flutter practice';
-  static const appVersion = 'Version: 1.1.1';
+  static const appVersion = 'Version: 1.1.2';
 }