+++ lib/logic/cubit/home_screen_cubit.dart
@@ -1,4 +1,5 @@
 import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 
 import '../../presentation/utils/custom_print.dart';
@@ -8,14 +9,9 @@ part 'home_screen_state.dart';
 class HomeScreenCubit extends Cubit<HomeScreenState> {
   HomeScreenCubit() : super(HomeScreenInitial());
 
-  void incrementer(){
-    customPrint.myCustomPrint('Inside cubit');
-    int newVal = 0;
+  void incrementer(int num){
+    customPrint.myCustomPrint(num+1);
 
-    if(state is HomeScreenIncremented){
-      newVal = (state as HomeScreenIncremented).value;
-    }
-
-    emit(HomeScreenIncremented(value: newVal+1));
+    emit(HomeScreenIncremented(value: num+1));
   }
 }